requirements.txt
amqp==5.1.0
apachelogs==0.6.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
dj-rest-auth==2.2.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
ipython==8.2.0
jedi==0.18.1
kombu==5.2.4
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydicti==1.1.6
Pygments==2.11.2
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
